Add feedTypes
