+++ src/scenes/tracksFeed/feedTypes.js
@@ -1,0 +1,2 @@
+export const GRID = 'grid';
+export const LIST = 'list';
